Rework API, add manual evaluation to endpoint
API now versioned, starting with v1.
Evaluation endpoint now POST rather than GET.
Evaluation endpoint now allows actual scaling, rather
than a read-only operation.
Read-only evaluation can be used with optional
dry_run query parameter.
