+++ api/v1/api.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 The Custom Pod Autoscaler Authors.
+Copyright 2020 The Custom Pod Autoscaler Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,25 +14,36 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package api provides routing and endpoints for the Custom Pod Autoscaler
-// HTTP REST API. Endpoints implemented as handlers, errors returned as valid
-// JSON.
-package api
+// Package v1 provides routing and endpoints for the Custom Pod Autoscaler
+// HTTP REST API version 1. Endpoints implemented as handlers, errors returned as
+// valid JSON.
+package v1
 
 import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+	"strconv"
 
 	"github.com/go-chi/chi"
 	"github.com/jthomperoo/custom-pod-autoscaler/config"
 	"github.com/jthomperoo/custom-pod-autoscaler/evaluate"
 	"github.com/jthomperoo/custom-pod-autoscaler/metric"
 	"github.com/jthomperoo/custom-pod-autoscaler/resourceclient"
+	"github.com/jthomperoo/custom-pod-autoscaler/scale"
 )
 
-// RunType api marks the metric gathering/evaluation as running during an API request
-const RunType = "api"
+const (
+	// RunType api marks the metric gathering/evaluation as running during an API request, which will use the results to scale
+	RunType = "api"
+	// RunTypeDryRun api marks the metric gathering/evaluation as running during an API request,
+	// which will only view the results and not use it for scaling
+	RunTypeDryRun = "api_dry_run"
+)
+
+const (
+	dryRunQueryParam = "dry_run"
+)
 
 // Error is an error response from the API, with the status code and an error message
 type Error struct {
@@ -45,6 +56,7 @@ type API struct {
 	Router          chi.Router
 	Config          *config.Config
 	Client          resourceclient.Client
+	Scaler          scale.Scaler
 	GetMetricer     metric.GetMetricer
 	GetEvaluationer evaluate.GetEvaluationer
 }
@@ -52,10 +64,12 @@ type API struct {
 // Routes sets up routing for the API
 func (api *API) Routes() {
 	// Set up routing
-	api.Router.Get("/metrics", api.getMetrics)
-	api.Router.Get("/evaluation", api.getEvaluation)
-	api.Router.NotFound(api.notFound)
-	api.Router.MethodNotAllowed(api.methodNotAllowed)
+	api.Router.Route("/api/v1", func(r chi.Router) {
+		r.NotFound(api.notFound)
+		r.MethodNotAllowed(api.methodNotAllowed)
+		r.Get("/metrics", api.getMetrics)
+		r.Post("/evaluation", api.getEvaluation)
+	})
 }
 
 func (api *API) getMetrics(w http.ResponseWriter, r *http.Request) {
@@ -90,6 +104,23 @@ func (api *API) getMetrics(w http.ResponseWriter, r *http.Request) {
 }
 
 func (api *API) getEvaluation(w http.ResponseWriter, r *http.Request) {
+	// Determine if it is a dry run
+	dryRun := true
+	dryRunParam := r.URL.Query().Get(dryRunQueryParam)
+	if dryRunParam == "" {
+		dryRun = false
+	} else {
+		b, err := strconv.ParseBool(dryRunParam)
+		if err != nil {
+			apiError(w, &Error{
+				fmt.Sprintf("Invalid format for 'dry_run' query parameter; '%s' is not a valid boolean value", dryRunParam),
+				http.StatusBadRequest,
+			})
+			return
+		}
+		dryRun = b
+	}
+
 	// Get resource being managed
 	resource, err := api.Client.Get(api.Config.ScaleTargetRef.APIVersion, api.Config.ScaleTargetRef.Kind, api.Config.ScaleTargetRef.Name, api.Config.Namespace)
 	if err != nil {
@@ -109,9 +140,15 @@ func (api *API) getEvaluation(w http.ResponseWriter, r *http.Request) {
 		})
 		return
 	}
+
+	// Set run type
 	metrics.RunType = RunType
+	if dryRun {
+		metrics.RunType = RunTypeDryRun
+	}
+
 	// Get evaluations for metrics
-	evaluations, err := api.GetEvaluationer.GetEvaluation(metrics)
+	evaluation, err := api.GetEvaluationer.GetEvaluation(metrics)
 	if err != nil {
 		apiError(w, &Error{
 			err.Error(),
@@ -119,8 +156,21 @@ func (api *API) getEvaluation(w http.ResponseWriter, r *http.Request) {
 		})
 		return
 	}
-	// Convert evaluations into JSON
-	response, err := json.Marshal(evaluations)
+
+	// Scale if not a dry run
+	if !dryRun {
+		evaluation, err = api.Scaler.Scale(*evaluation, resource, api.Config.MinReplicas, api.Config.MaxReplicas, api.Config.ScaleTargetRef, api.Config.Namespace)
+		if err != nil {
+			apiError(w, &Error{
+				err.Error(),
+				http.StatusInternalServerError,
+			})
+			return
+		}
+	}
+
+	// Convert evaluation into JSON
+	response, err := json.Marshal(evaluation)
 	if err != nil {
 		// Should not occur, panic
 		panic(err)
