+++ example/simple-pod-metrics-python/Dockerfile
@@ -18,8 +18,8 @@ FROM custompodautoscaler/python:latest
 RUN apt-get update
 RUN apt-get install jq curl -y
 # Set up aliases for easy debugging
-RUN echo 'alias metrics="curl http://localhost:5000/metrics | jq ."' >> ~/.bashrc
-RUN echo 'alias evaluation="curl http://localhost:5000/evaluation | jq ."' >> ~/.bashrc
+RUN echo 'alias metrics="curl http://localhost:5000/api/v1/metrics | jq ."' >> ~/.bashrc
+RUN echo 'alias evaluation="curl http://localhost:5000/api/v1/evaluation | jq ."' >> ~/.bashrc
 # Add in python requirements to install dependencies
 ADD requirements.txt /
 # Install dependencies, done before adding Py files to allow Docker to cache better
